Saves show/hide read status.
This is an alternative implementation of the feature of commits
  50f9b2088e04e1b782d72aaab7fc90ff26906a40
    (w/ fix af35c5b57eac113a267e9f9c721b69a114d9db79)
  b19ec1b7df4f11604841312efab56345c16730e1
  06b759d7afa731c1d157a5d57939241b3a217439
from Mathias Weinerts repository at https://code.google.com/r/matweinert-sparserss/

In this implementation, a database field is used for 'real' feeds such that the
setting gets deleted if the feed gets deleted. As there is the need for a new
db field, the logic of 'show_read' has been inverted to 'hide_read'. The
setting can be also changed at the edit-feed-dialog. Just as in Mathias'
version, the setting for the 'virtual' feeds (all entries and favorites) is stored
in preferences.

This also closes issue 74 (Global setting for "hide read entries"). However, the
suggestion there was to add a global setting. Instead, a per feed setting is
implemented.