Fetch feed icon url from html head tag.
Based on the regex of commit 0f9e78740aae011ead1d769af4259ee267135b71 from https://code.google.com/r/matweinert-sparserss/
by Mathias Weinert, this adds a different implementation. The head-tag is
already parsed while trying to find the proper feed link (out of the website)
and therefore, the icon should be searched for in this part.
Additionally, the regex is case-insensitive which should make it more
successful in finding icons.